requirements.txt
flask
numpy
paho.mqtt.client
pandas